+++ Runtime/Compressors/ICompressor.cs
@@ -1,8 +1,16 @@
+using System.Threading.Tasks;
+
 namespace ActionCode.Persistence
 {
     public interface ICompressor
     {
-        string Compress(string value);
+        /// <summary>
+        /// Compress the given value.
+        /// </summary>
+        /// <param name="value">A string to compress.</param>
+        /// <returns>A task operation of the compress process.</returns>
+        Task<string> Compress(string value);
+
         string Decompress(string value);
     }
 }